Move Siae model to own app
