+++ lemarche/siaes/validators.py
@@ -1,0 +1,11 @@
+from django.core.exceptions import ValidationError
+
+
+def validate_siret(siret):
+    if not siret.isdigit() or len(siret) != 14:
+        raise ValidationError("Le numéro SIRET doit être composé de 14 chiffres.")
+
+
+def validate_naf(naf):
+    if len(naf) != 5 or not naf[:4].isdigit() or not naf[4].isalpha():
+        raise ValidationError("Le code NAF doit être composé de de 4 chiffres et d'une lettre.")
